move all the js to a single file
